% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartbookR.R
\docType{package}
\name{chartbookR}
\alias{chartbookR}
\alias{chartbookR-package}
\title{Creation of chartbooks}
\description{
chartbookR: Offers convenient download and transformation of Bloomberg and FRED (St. Louis Fed) data, plotting of customized charts, and saving to Powerpoint.
}
\section{Details}{

The chartbookR package consists of four types of functions for chartbook creation:

\itemize{
\item Data download: \code{\link[chartbookR]{getBBG}} and \code{\link{getFRED}}
\item Data transformation: \code{\link{Transform}}
\item Customized plotting: \code{\link{LineChart}} and \code{\link{BarChart}}
\item Exporting to Powerpoint: \code{\link{makePPT}}
}
}

\section{Notes}{

While all functions are based on existing packages, they make the chartbook creation an efficient and convenient process.

In order to use the \code{\link{getBBG}} function, you need to have a running Bloomberg terminal session. If this is not the case, you can still use \code{\link{getFRED}} to download freely available data from the St. Louis Fed database (\url{https://fred.stlouisfed.org}).
}

\section{Author/Maintainer}{

Kilian Reber \email{kilian.d.reber@gmail.com}
}

\section{Contributors}{

Tryggvi Thoroddsen \email{tryggvithoroddsen@gmail.com}
}

\section{See Also}{

\url{https://github.com/kilianreber}
}

