% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBBG.R
\name{getBBG}
\alias{getBBG}
\title{Bloomberg data as zoo object}
\usage{
getBBG(tickers, field, names, time, start, end, freq, na)
}
\arguments{
\item{tickers}{character vector of Bloomberg tickers}

\item{field}{optional character vector of one or several Bloomberg fields; default is 'PX_LAST'}

\item{names}{optional character vector of column names for zoo object; default uses 'tickers' vector}

\item{time}{optional string indicating the lookback from today; options are 'D' (Days), 'W' (Weeks), 'M' (Months), 'Q' (Quarters), 'Y' (Years), or 'YTD' (Year-to-Date), e.g. '3M', '4Q', '5Y', 'YTD'; default is none}

\item{start}{optional start date for data download; default is Sys.Date() - 3*365}

\item{end}{optional end date for data download; default is Sys.Date() -1}

\item{freq}{optional frequency for data download; options are 'DAILY', 'WEEKLY', 'MONTHLY', 'QUARTERLY', 'YEARLY'; default is 'MONTHLY'}

\item{na}{optional boolean which will fill intermittent NAs if set to TRUE; default is FALSE}
}
\value{
returns a zoo object with the downloaded Bloomberg data
}
\description{
Returns Bloomberg data as zoo object
}
\examples{
df <- getBBG(tickers="VIX Index", time="YTD")
df <- getBBG(tickers="VIX Index", names="VIX", freq="D", time="30Y", na=FALSE)
df <- getBBG(tickers=c("CPI YOY Index", "PPI YOY Index"), names=c("CPI", "PPI"), start=as.Date("2000-01-01"))
df <- getBBG(tickers=c("NAPMPMI", "MPMIEZMA", "MPMIEMMA"), names=c("United States (ISM)", "Eurozone", "Emerging Markets"), time="3Y")
}
