% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyChart.R
\name{HyChart}
\alias{HyChart}
\title{Returns highcharter object}
\usage{
HyChart(data, l1 = NULL, l2 = NULL, a1 = NULL, a2 = NULL,
  c1 = NULL, c2 = NULL, series1 = NULL, stacking = NULL,
  title = NULL, subtitle = NULL, title_adj = "left", y1 = "",
  y2 = "", y1_def = c(NULL, NULL, NULL), y2_def = c(NULL, NULL,
  NULL), y1_right = FALSE, y1_rev = FALSE, y2_rev = FALSE,
  grid = c(1, 1), lineWidth = 3, zoom = TRUE, zoom_cst = NULL,
  navigator = FALSE, tooltip = TRUE, decimals = 1, size = c(750,
  600), linePos = 1, h1 = NULL, h2 = NULL, v = NULL,
  b1_from = NULL, b1_to = NULL, b2_from = NULL, b2_to = NULL,
  v_from = NULL, v_to = NULL, rec = FALSE)
}
\description{
Returns highcharter object
}
