% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineChart.R
\name{LineChart}
\alias{LineChart}
\title{Create custom line(s) plot}
\usage{
LineChart(data, inception = FALSE, title = NULL, title_adj = NULL,
  no = "", l1, l2 = "none", y1 = NULL, y2 = NULL,
  y1_def = "none", y2_def = "none", y2_rev = FALSE, fn = NULL,
  fn_adj = NULL, leg = "topleft", grid = FALSE, rec = FALSE,
  dt_format, h1 = "none", v = "none")
}
\arguments{
\item{data}{specification of zoo dataset to use for plot}

\item{inception}{optional boolean to show clipped inception date when set to 'TRUE'}

\item{title}{optional character to add chart title}

\item{title_adj}{optional integer to specify title alignment (0: left-align, 1: right-align, NULL: center)}

\item{no}{optional integer to add chart number}

\item{l1}{optional integer vector to specify which columns to plot (as lines) on primary y-axis}

\item{l2}{optional integer vector to specify which columns to plot (as lines) on secondary y-axis}

\item{y1}{optional character to specify description of primary y-axis}

\item{y2}{optional character to specify description of secondary y-axis}

\item{y1_def}{optional number vector to specify start, end, and intervals of primary y-axis, e.g. c(0, 10, 2)}

\item{y2_def}{optional number vector to specify start, end, and intervals of secondary y-axis, e.g. c(0, 10, 2)}

\item{y2_rev}{optional boolean to invert secondary y-axis; y2_def needs to be supplied}

\item{fn}{optional character to add footnote to plot}

\item{fn_adj}{optional integer to specify alignment of footnote (0: left-align, 1: right-align, NULL: center)}

\item{leg}{optional character to specify legend position; options are 'topleft', 'center', 'topright', 'left', 'center', 'right', 'bottomleft', 'bottom', 'bottomright'}

\item{grid}{optional boolean to show grid when set to TRUE}

\item{rec}{optional boolean to shade recessions when set to TRUE}

\item{dt_format}{optional character vector to specify date interval and date format of x-axis, e..g dt_format=c('3 months', '\%b-\%Y')}

\item{h1}{optional argument to specify horizontal line(s) at specific height}

\item{v}{optional date to specify vertical date line(s), e.g. v='31/12/2012'}
}
\value{
plots R base plot
}
\description{
Create custom line(s) plot based on user inputs
}
\examples{

Note that these examples require corresponding zoo objects to work properly

LineChart(data = zoo, l1 = 1:3, leg = "topleft", fn = "This is a footnote.")
LineChart(data = zoo, title = "Example Chart", no = "1.1", l1 = 1:3, y1 = "Index", leg = "topleft", dt_format = c("5 years", "\%Y"), v = "01/01/2018")
LineChart(data = zoo, title = "Example Chart", l1 = 1:3, y1 = "Index", leg = "topleft", grid = TRUE, h = 0)
LineChart(data = zoo, title = "Example Chart", l1 = 4:5, l2 = 6, y1 = "in \%", y2 = "in USD mln", rec = TRUE)
LineChart(data = zoo, title = "Example Chart", l1 = 1, l2 = 2, y1 = "Unemployment (\%)", y2 = "Budget Balance (\%GDP)", y1_def = c(0, 10, 2), y2_def = c(-10,2,2), y2_rev = TRUE, leg = "top", rec = TRUE)
}
