% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPT.R
\name{makePPT}
\alias{makePPT}
\title{Create Powerpoint slide from R plot viewer}
\usage{
makePPT(path, width, height, left, top, header, source, source_dt)
}
\arguments{
\item{path}{required string to specify path for plot viewer temp file}

\item{width}{optional integer to specify chart width}

\item{height}{optional integer to specify chart height}

\item{left}{optional integer to specify left-hand margins on Powerpoint slide}

\item{header}{optional character string to specify Powerpoint slide header}

\item{source}{optional character string to specify Powerpoint slide source note}

\item{source_dt}{optional character string to specify Powerpoint slide source date}

\item{left}{optional integer to specify top margins on Powerpoint slide}
}
\value{
Adds plot viewer content to Powerpoint slide
}
\description{
Adds plot viewer content to Powerpoint slide; it relies on the package 'officer'
}
\examples{
Before charts can be added to a Powerpoint file, it needs to be initiated using a pptx Template file via: 
doc <- read_pptx("C:/Documents/Powerpoint/Template.pptx")

Charts can then be added via: 
makePPT(path="C:/temp", header="US GDP Figures", source="Bloomberg, Fed. Saint Louis")

Once all charts are added, the Powerpoint file is saved via: 
print(doc, target="C:/Documents/Powerpoint/US_Macro_Slideset.pptx")
}
