% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform.r
\name{Transform}
\alias{Transform}
\title{Calculate transformation of zoo object}
\usage{
Transform(data, start, chg, chg_type, pma, pms, lag, lead, rebase, Z)
}
\arguments{
\item{data}{specification of zoo object}

\item{start}{optional start date to trim 'data'}

\item{chg}{optional string to calculate periodical changes; options are 'YOY' (Year-on-Year), 'QOQ' (Quarter-on-Quarter), 'MOM' (Month-on-Month); is appended for all columns}

\item{chg_type}{optional string to specify calculation of changes; options are 'perc' (percentage change) or 'delta' (absolute change); default is 'perc'}

\item{pma}{optional integer to specify calculation of period-moving-averages; is appended for all columns}

\item{pms}{optional integer to specify calculation of period-moving-sums; is appended for all columns}

\item{lag}{optional integer to lag the data; is appended for all columns}

\item{lead}{optional integer to lead the data; is appended for all columns}

\item{rebase}{optional boolean to rebase the data to 100 at first observation; is appended for all columns}

\item{Z}{optional integer to specify the rolling number of periods over which the data is normalized; is appended for all columns}
}
\value{
returns a zoo object with transformations appended
}
\description{
Transforms a zoo object and appends results
}
