% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBBG.R
\name{getBBG}
\alias{getBBG}
\title{Bloomberg data as zoo object}
\usage{
getBBG(tickers, field, names, start, end, time, freq, na)
}
\arguments{
\item{tickers}{character vector of Bloomberg ticker(s)}

\item{field}{optional character vector of Bloomberg field(s); default is 'PX_LAST'}

\item{names}{optional character vector of column name(s) for zoo object; default uses 'tickers' vector}

\item{start}{optional start date for data download; default is Sys.Date() - 3*365}

\item{end}{optional end date for data download; default is Sys.Date() -1}

\item{time}{optional string to specify start date; options are 'D' (Days), 'W' (Weeks), 'M' (Months), 'Q' (Quarters), 'Y' (Years), or 'YTD' (Year-to-Date), e.g. '3M', '4Q', '5Y', 'YTD'; default is none}

\item{freq}{optional frequency for data download; options are 'DAILY', 'WEEKLY', 'MONTHLY', 'QUARTERLY', 'YEARLY'; default is 'MONTHLY'}

\item{na}{optional boolean to replace NAs with the last observation if set to FALSE; default is TRUE}
}
\value{
returns a zoo object with the downloaded Bloomberg data
}
\description{
Returns Bloomberg data as zoo object; it relies on Rblpapi package but simplifies download and output
}
\examples{
zoo <- getBBG(tickers='VIX Index', time='YTD')
zoo <- getBBG(tickers='VIX Index', names='VIX', freq='D', time='30Y', na=FALSE)
zoo <- getBBG(tickers=c('CPI YOY Index', 'PPI YOY Index'), names=c('CPI', 'PPI'), start='2000-01-01')
zoo <- getBBG(tickers=c('CPI YOY Index', 'PPI YOY Index'), names=c('CPI', 'PPI'), start='2000-01-01', end='2018-01-01')
zoo <- getBBG(tickers=c('NAPMPMI', 'MPMIEZMA', 'MPMIEMMA'), names=c('United States (ISM)', 'Eurozone', 'Emerging Markets'), time='3Y')
}
