% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFRED.R
\name{getFRED}
\alias{getFRED}
\title{FRED data as zoo object}
\usage{
getFRED(tickers, names, start, end, time, na)
}
\arguments{
\item{tickers}{character vector of FRED ticker(s)}

\item{names}{optional character vector of column name(s) for zoo object; default uses 'tickers' vector}

\item{start}{optional start date for data download; default is Sys.Date() - 3*365}

\item{end}{optional end date for data download; default is Sys.Date() -1}

\item{time}{optional string to specify start date; options are 'D' (Days), 'W' (Weeks), 'M' (Months), 'Q' (Quarters), 'Y' (Years), or 'YTD' (Year-to-Date), e.g. '3M', '4Q', '5Y', 'YTD'; default is none}

\item{na}{optional boolean to fill intermittent NAs if set to FALSE; default is TRUE}
}
\value{
returns a zoo object with the downloaded FRED data
}
\description{
Returns FRED data as zoo object; it relies on quantmod::getSymbols but simplifies download and output
}
\examples{
df <- getFRED(tickers='CPIAUCSL', time='CPI')
df <- getFRED(tickers='CPIAUCSL', names='CPI', time='30Y', na=FALSE)
df <- getFRED(tickers=c("CPIAUCSL", "PCEPILFE"), names=c("CPI", "Core PCE"), start='2000-01-01')
df <- getFRED(tickers=c("CPIAUCSL", "PCEPILFE"), names=c("CPI", "Core PCE"), start='2000-01-01', end='2018-01-01')
df <- getFRED(tickers=c("UNRATE", "PCEPILFE"), names=c("Unemployment Rate", "Core PCE"), time='3Y')
}
